==> 0-preprocess-generate_csvs.ipynb <==
import pandas as pd
from pathlib import Path
from tqdm import tqdm_notebook
import pydicom
import itertools
import numpy as np
from concurrent.futures import ProcessPoolExecutor

%%capture
from tqdm import tqdm_notebook as tqdm
tqdm().pandas()

# Read stage_X_train and split id/label

stage = "stage_2"

bad_dcm_fn = f'data/unzip/{stage}_train_images/ID_6431af929.dcm'

!rm {bad_dcm_fn}

df_train = pd.read_csv(f'data/unzip/{stage}_train.csv')

df_train.head()

df_train['fid'] = df_train.ID.apply(lambda x: '_'.join(x.split('_')[:2]) )

df_train.columns = ['ID', 'probability', 'fid']

df_train['label'] = df_train.ID.apply(lambda x: x.split('_')[-1])

df_train.head()

# Remove dupes 

df_train.shape

df_train.drop_duplicates('ID', inplace=True)

df_train.shape

# Remove corrupted image

df_train = df_train[df_train.fid != 'ID_6431af929'] # ID_6431af929

df_train.shape

# Create pivot table with diagnostic labels as columns
Generates:
* `train_diags.csv` (previously named `train_pivot.csv`)

df_diags = df_train.pivot(index='fid', columns='label', values='probability')

df_diags.head()

df_diags.shape

df_diags.reset_index(inplace=True)

df_diags.head()

df_diags.shape

df_diags.to_csv(f'data/{stage}_train_diags.csv', index=False)

# Generate fastai-ready csv image file (.png) -> labels
This is needed for early experiments that worked with the .png dataset.

Generates:
* `train_labels_as_strings.csv`

from collections import defaultdict

d = defaultdict(list)
for fid in df_train.fid.unique(): d[fid]

for tup in df_train.itertuples():
    if tup.probability: d[tup.fid].append(tup.label)

ks, vs = [], []

for k, v in d.items():
    ks.append(k), vs.append(' '.join(v))

fastai_df = pd.DataFrame(data={'fn': ks, 'labels': vs})

fastai_df.shape

fastai_df['fn'] += '.png'

fastai_df.head()

fastai_df.to_csv(f'data/{stage}_train_labels_as_strings.csv', index=False)

# Tabulate dicom data
Generates:
* `train_dicom.csv`
* `test_dicom.csv`

cols_i_want = ['BitsAllocated', 'BitsStored', 'Columns', 'HighBit', 'ImageOrientationPatient', 
               'ImagePositionPatient', 'Modality', 'PatientID', 'PhotometricInterpretation', 
               'PixelRepresentation', 'PixelSpacing', 'RescaleIntercept', 'RescaleSlope', 
               'Rows', 'SOPInstanceUID', 'SamplesPerPixel', 'SeriesInstanceUID', 'StudyID', 
               'StudyInstanceUID', 'WindowCenter', 'WindowWidth']
useless_cols = [ 'PixelData' ]

train_dcm_list = list(Path(f'data/unzip/{stage}_train_images').glob('*.dcm'))

f = train_dcm_list[0]

dicom = pydicom.dcmread(str(f))
dicom

dicom.data_element('PatientID').value

ipp = dicom.data_element('ImagePositionPatient').value
ipp

ipp[0]

type(ipp)

def tabulate_dicom_data(file_list):
    l = []
    for f in file_list:
        dicom = pydicom.dcmread(str(f))
        d = {}
        for s in cols_i_want:
            v = dicom.data_element(s).value
            if isinstance(v, pydicom.multival.MultiValue):
                for i in range(len(v)):
                    d[s + '_' + str(i)] = v[i]
            else:
                d[s] = v
        l.append(d)
        
    return l


with ProcessPoolExecutor(max_workers=32) as e:
     l = list(itertools.chain.from_iterable(e.map(tabulate_dicom_data, np.array_split(train_dcm_list, 32))))

df_train_dicom = pd.DataFrame(l)

df_train_dicom.head()

df_train_dicom.to_csv(f'data/{stage}_train_dicom.csv')

test_dcm_list = list(Path(f'data/unzip/{stage}_test_images').glob('*.dcm'))
with ProcessPoolExecutor(max_workers=32) as e:
     l = list(itertools.chain.from_iterable(e.map(tabulate_dicom_data, np.array_split(test_dcm_list, 32))))

df_test_dicom = pd.DataFrame(l)
df_test_dicom.head()

df_test_dicom.to_csv(f'data/{stage}_test_dicom.csv')

# Add labels to the train dicom csv
Generates:
* `train_dicom_diags.csv` (previously named `train_dicom_pivot.csv`)

df_train_dicom_diags = pd.merge(df_train_dicom, df_diags,  how='left', left_on=['SOPInstanceUID'], right_on = ['fid'])

assert len(df_train_dicom) == len(df_diags) == len(df_train_dicom_diags)

df_train_dicom_diags.to_csv(f'data/{stage}_train_dicom_diags.csv')
