==> 2-preprocess-pickle.ipynb <==
import pandas as pd
from pathlib import Path
from collections import defaultdict
import pickle

stage = "stage_2"

# Train dataset

df = pd.read_csv(f"data/{stage}_train_dicom_diags.csv")

df.head()

# sort, then group by (order is preserved within groups)
gs = df.sort_values('ImagePositionPatient_2').groupby('SeriesInstanceUID')
len(gs)

# see if it worked
gs.get_group('ID_fa19cd5ea9')[['ImagePositionPatient_2', 'fid']].head()

g = gs.get_group('ID_fa19cd5ea9')

subg = g[['SeriesInstanceUID', 'fid', 'any', 'epidural', 
          'intraparenchymal', 'intraventricular', 'subarachnoid', 'subdural']]

subg

# You can use a btrfs snapshot and rename files by study_id and z-pos through the brain
def rename_train_group(subg):
    ix = 0
    total = len(subg)
    for index, row in subg.iterrows():
        cur_fn = row['fid']
        new_fn = f"{row['SeriesInstanceUID']}_{ix:03}_{total:03}_{row['any']}_{row['epidural']}_{row['intraparenchymal']}_{row['intraventricular']}_{row['subarachnoid']}_{row['subdural']}_{cur_fn}"
        ix += 1
        Path(f'data/unzip_renamed/{stage}_train_images/{cur_fn}.dcm').rename(f'data/unzip_renamed/{stage}_train_images/{new_fn}.dcm')

def index_group(subg, study_ix_to_fn, fn_to_study_ix):
    ix = 0
    for index, row in subg.iterrows():
        fn = row['SOPInstanceUID']
        study = row['SeriesInstanceUID']
        study_ix_to_fn[study].append(fn)
        fn_to_study_ix[fn] = (study, ix)
        ix += 1  

labels = [ 'any', 'epidural', 'intraparenchymal', 'intraventricular', 'subarachnoid', 'subdural' ]

def label_group(subg, fn_to_labels):
    for index, row in subg.iterrows():
        fn = row['SOPInstanceUID']
        fn_to_labels[fn] = [ label for label in labels if row[label] == 1 ]

train_study_ix_to_fn = defaultdict(list)
train_fn_to_study_ix = {}
train_fn_to_labels = {}

for name, subg in gs:
    #rename_train_group(subg)
    index_group(subg, train_study_ix_to_fn, train_fn_to_study_ix)
    label_group(subg, train_fn_to_labels)

# Do not pickle yet, merge with test
pickle.dump(train_study_ix_to_fn, open(f"data/{stage}_train_study_ix_to_fn.pickle", "wb" ))
pickle.dump(train_fn_to_study_ix, open(f"data/{stage}_train_fn_to_study_ix.pickle", "wb" ))

train_fn_to_labels['ID_079945c27']

pickle.dump(train_fn_to_labels, open(f"data/{stage}_train_fn_to_labels.pickle", 'wb'))

# Test dataset

df = pd.read_csv(f"data/{stage}_test_dicom.csv")

df.head()

# sort, then group by (order is preserver within groups)
gs = df.sort_values('ImagePositionPatient_2').groupby('SeriesInstanceUID')
len(gs)

def rename_test_group(subg):
    ix = 0
    total = len(subg)
    for index, row in subg.iterrows():
        cur_fn = row['SOPInstanceUID']
        new_fn = f"{row['SeriesInstanceUID']}_{ix:03}_{total:03}_{cur_fn}"
        ix += 1
        Path(f'data/unzip_renamed/{stage}_test_images/{cur_fn}.dcm').rename(f'data/unzip_renamed/{stage}_test_images/{new_fn}.dcm')


test_study_ix_to_fn = defaultdict(list)
test_fn_to_study_ix = {}

for name, subg in gs:
    #rename_test_group(subg)
    index_group(subg, test_study_ix_to_fn, test_fn_to_study_ix)


pickle.dump(test_study_ix_to_fn, open(f"data/{stage}_test_study_ix_to_fn.pickle", "wb" ))
pickle.dump(test_fn_to_study_ix, open(f"data/{stage}_test_fn_to_study_ix.pickle", "wb" ))

study_ix_to_fn = { **train_study_ix_to_fn, **test_study_ix_to_fn }
fn_to_study_ix = { **train_fn_to_study_ix, **test_fn_to_study_ix }

pickle.dump(study_ix_to_fn, open(f"data/{stage}_study_ix_to_fn.pickle", "wb" ))
pickle.dump(fn_to_study_ix, open(f"data/{stage}_fn_to_study_ix.pickle", "wb" ))
