==> 2a-preprocess-pickle-study_to_data.ipynb <==
import fastai
import pickle
from fastai.utils.mem import *

stage = 'stage_2'
data_dir = Path('data')
fn_to_study_ix = pickle.load(open(f'{data_dir}/{stage}_fn_to_study_ix.pickle', 'rb'))
study_ix_to_fn = pickle.load(open(f'{data_dir}/{stage}_study_ix_to_fn.pickle', 'rb'))
fn_to_labels = pickle.load(  open(f'{data_dir}/{stage}_train_fn_to_labels.pickle', 'rb'))

classes = ['any', 'epidural', 'intraparenchymal', 'intraventricular', 'subarachnoid', 'subdural']
labels = []
for fn, lbls in fn_to_labels.items():
    row = {'fn':fn,'study':fn_to_study_ix[fn][0]}
    for c in classes:
        row[c] = c in fn_to_labels[fn]
    labels.append(row)

labels_df = pd.DataFrame(labels)
labels_df.head() 

study_labels = labels_df.groupby('study').agg('sum').reindex()
for c in classes :
    study_labels[c+'_size'] = 0 # 0 :not present, 1:small, 2:big
    c_idx = study_labels.query(f'{c}>0').sort_values(c).index
    c_small, c_big = c_idx[:len(c_idx)//2], c_idx[len(c_idx)//2:]
    study_labels.loc[c_small,c+'_size'] = 1
    study_labels.loc[c_big,c+'_size'] = 2
size_classes = [c+'_size' for c in classes]
study_labels['strat_class'] = study_labels[size_classes].astype(str).sum(axis=1)
study_labels.sample(10)

from sklearn.model_selection import StratifiedKFold

#this will throw a warning that can be ignored
skf = StratifiedKFold(n_splits=19, shuffle=True, random_state=1972)
study_labels['fold'] = -1
for fold, (oof_idx,f_idx) in enumerate(skf.split(study_labels, study_labels.strat_class)):
    study_labels.loc[study_labels.iloc[f_idx].index, 'fold' ] = fold
study_labels.head()

study_to_data = {}
for study in study_ix_to_fn.keys():
    if study in study_labels.index:
        study_to_data[study] = {'fold': study_labels.loc[study].fold}
    else:
        study_to_data[study] = {'fold': -1} #study not in label set
pickle.dump(study_to_data, open(f"data/{stage}_study_to_data.pickle", "wb" ))
